try seperating logic from ui as much as I can; try modify jacoco to exclude ui folder
